requirements.txt
streamlit
numpy
pillow
tensorflow
gdown
